Add cisagov/ansible-role-debian-archive
Debian Buster is no longer supported by the standard Debian package
repository URLs because it is so outdated; therefore, this platform
must be switched to use the Debian Archive package package
repositories.

Note that the cisagov/ansible-role-debian-archive role must be
applied *before* cisagof/ansible-role-upgrade, so the role *must* be
added to the upgrade.yml playbook.